Define Fifo::size_type as size_t.
Elsewhere in the code we invoke "std::min" on a Fifo::size_type and a
size_t.  size_type was an unsigned int, which is less common than size_t
for this purpose.

	* include/libport/fifo.hh: here.

+++ include/libport/fifo.hh
@@ -48,7 +48,7 @@ namespace libport
     typedef typename Alloc::pointer pointer;
     typedef typename Alloc::const_reference const_reference;
     typedef typename Alloc::const_pointer const_pointer;
-    typedef unsigned int size_type;
+    typedef size_t size_type;
     typedef Alloc allocator_type;
 
     //! Fifo constructor.
